fix: Entering a folder on the hard drive and unplugging the storage device, it was found that when the dde-file-manager displayed "empty folder" and the device was inserted, the folder was opened again,and the entire file was unknown
After uninstallation, the fileinfo created on the USB drive for subsequent processing is set to local, while the fileinfo created on another directory is an unknown attribute. In the second iteration, it is determined whether the fileinfo is local, resulting in an error that causes the iteration to go to local. Therefore, it is not necessary to obtain the fileinfo in the iteration. Which directory's unknown fileinfo was obtained

log: show error file icon
bug: https://pms.uniontech.com/bug-view-276311.html https://pms.uniontech.com/bug-view-279091.html
Change-Id: Iba6899b1132659f6615ba2f71a67d2cc851a8290