fix: Prevent fd leak when copying files from a disc
An iterator that opens the disc's mountpoint was not properly closing
its file descriptor during the copy operation.

This resource leak caused an issue where, if a disc was ejected mid-copy
and the tray was subsequently closed, `udev` would fail to obtain the
new disc's information because the fd for the original mountpoint was
still held open.

This change ensures the file descriptor is closed correctly, resolving
the issue.

See: https://pms.uniontech.com/bug-view-323377.html
See: https://pms.uniontech.com/bug-view-335269.html

Bug: https://pms.uniontech.com/bug-view-335271.html
Change-Id: I0f185b7eac01c22e03453ecb169990dabc56d215